+++ firmware/controllers/algo/airmass/maf_airmass.cpp
@@ -31,7 +31,7 @@ float MafAirmass::getMaf() const {
 	}
 }
 
-AirmassResult MafAirmass::getAirmass(int rpm, bool postState) {
+AirmassResult MafAirmass::getAirmass(float rpm, bool postState) {
 	float maf = getMaf();
 
 	return getAirmassImpl(maf, rpm, postState);
@@ -41,7 +41,7 @@ AirmassResult MafAirmass::getAirmass(int rpm, bool postState) {
  * Function block now works to create a standardised load from the cylinder filling as well as tune fuel via VE table. 
  * @return total duration of fuel injection per engine cycle, in milliseconds
  */
-AirmassResult MafAirmass::getAirmassImpl(float massAirFlow, int rpm, bool postState) const {
+AirmassResult MafAirmass::getAirmassImpl(float massAirFlow, float rpm, bool postState) const {
 	// If the engine is stopped, MAF is meaningless
 	if (rpm == 0) {
 		return {};
