+++ firmware/controllers/algo/fuel_math.h
@@ -26,7 +26,7 @@ float getCltFuelCorrection();
 angle_t getCltTimingCorrection();
 float getCrankingFuel(float baseFuel);
 float getCrankingFuel3(float baseFuel, uint32_t revolutionCounterSinceStart);
-float getInjectionMass(int rpm);
+float getInjectionMass(float rpm);
 percent_t getInjectorDutyCycle(int rpm);
 percent_t getInjectorDutyCycleStage2(int rpm);
 float getStage2InjectionFraction(int rpm, float fuelLoad);
